move geocoding functions to helpers
