+++ test_cli.py
@@ -4,7 +4,7 @@ from click.testing import CliRunner
 import responses
 
 from mapboxcli.scripts.cli import main_group
-from mapboxcli.scripts.geocoding import coords_from_query, iter_query
+from mapboxcli.scripts.helpers import coords_from_query, iter_query
 
 
 def test_iter_query_string():
